enhance MQTT connection handling, improve SysML model parsing, and standardize diagnostic outputs across components.

+++ DigitalTwinLib/src/Model/Entities/Package.cpp
@@ -63,7 +63,7 @@ namespace DigitalTwin::Model
 
 	Variable* Package::resolveVariable(std::vector<std::string> domains, int index)
 	{
-		if (index >= domains.size())
+		if (static_cast<size_t>(index) >= domains.size())
 			throw DigitalTwinAddressException();
 
 		if ((size_t)index == (domains.size() - 1))
